chore: add fee service class
