+++ packages/relay/src/lib/services/ethService/ethCommonService/index.ts
@@ -14,7 +14,7 @@ import {
   prepend0x,
   stripLeadingZeroForSignatures,
   tinybarsToWeibars,
-  toHash32,
+  toHash32
 } from '../../../../formatters';
 import { MirrorNodeClient } from '../../../clients';
 import constants from '../../../constants';
@@ -26,6 +26,9 @@ import { Log } from '../../../model';
 import { RequestDetails } from '../../../types';
 import { CacheService } from '../../cacheService/cacheService';
 import { ICommonService } from './ICommonService';
+import { Utils } from '../../../../utils';
+import { Hbar } from '@hashgraph/sdk';
+import HAPIService from '../../hapiService/hapiService';
 
 /**
  * Create a new Common Service implementation.
@@ -41,6 +44,13 @@ export class CommonService implements ICommonService {
    * @private
    */
   private readonly mirrorNodeClient: MirrorNodeClient;
+
+  /**
+   * The client service which is responsible for client all logic related to initialization, reinitialization and error/transactions tracking.
+   *
+   * @private
+   */
+  private readonly hapiService: HAPIService;
 
   /**
    * The logger used for logging all output from this class.
@@ -68,7 +78,11 @@ export class CommonService implements ICommonService {
   private readonly maxBlockRange = parseNumericEnvVar('MAX_BLOCK_RANGE', 'MAX_BLOCK_RANGE');
   private readonly ethBlockNumberCacheTtlMs = parseNumericEnvVar(
     'ETH_BLOCK_NUMBER_CACHE_TTL_MS',
-    'ETH_BLOCK_NUMBER_CACHE_TTL_MS_DEFAULT',
+    'ETH_BLOCK_NUMBER_CACHE_TTL_MS_DEFAULT'
+  );
+  private readonly ethGasPriceCacheTtlMs = parseNumericEnvVar(
+    'ETH_GET_GAS_PRICE_CACHE_TTL_MS',
+    'ETH_GET_GAS_PRICE_CACHE_TTL_MS_DEFAULT'
   );
 
   // Maximum allowed timestamp range for mirror node requests' timestamp parameter is 7 days (604800 seconds)
@@ -78,10 +92,11 @@ export class CommonService implements ICommonService {
     return ConfigService.get('ETH_GET_LOGS_BLOCK_RANGE_LIMIT');
   }
 
-  constructor(mirrorNodeClient: MirrorNodeClient, logger: Logger, cacheService: CacheService) {
+  constructor(mirrorNodeClient: MirrorNodeClient, logger: Logger, cacheService: CacheService, hapiService: HAPIService) {
     this.mirrorNodeClient = mirrorNodeClient;
     this.logger = logger;
     this.cacheService = cacheService;
+    this.hapiService = hapiService;
   }
 
   public static blockTagIsLatestOrPendingStrict(tag: string | null): boolean {
@@ -103,7 +118,7 @@ export class CommonService implements ICommonService {
     fromBlock: string,
     toBlock: string,
     requestDetails: RequestDetails,
-    address?: string | string[] | null,
+    address?: string | string[] | null
   ) {
     if (this.blockTagIsLatestOrPending(toBlock)) {
       toBlock = CommonService.blockLatest;
@@ -161,7 +176,7 @@ export class CommonService implements ICommonService {
           prepend0x(fromBlockNum.toString(16)),
           fromBlockResponse.timestamp.from,
           prepend0x(toBlockNum.toString(16)),
-          toBlockResponse.timestamp.to,
+          toBlockResponse.timestamp.to
         );
       }
 
@@ -241,12 +256,12 @@ export class CommonService implements ICommonService {
   public async getHistoricalBlockResponse(
     requestDetails: RequestDetails,
     blockNumberOrTagOrHash?: string | null,
-    returnLatest?: boolean,
+    returnLatest?: boolean
   ): Promise<any> {
     if (!returnLatest && this.blockTagIsLatestOrPending(blockNumberOrTagOrHash)) {
       if (this.logger.isLevelEnabled('debug')) {
         this.logger.debug(
-          `${requestDetails.formattedRequestId} Detected a contradiction between blockNumberOrTagOrHash and returnLatest. The request does not target the latest block, yet blockNumberOrTagOrHash representing latest or pending: returnLatest=${returnLatest}, blockNumberOrTagOrHash=${blockNumberOrTagOrHash}`,
+          `${requestDetails.formattedRequestId} Detected a contradiction between blockNumberOrTagOrHash and returnLatest. The request does not target the latest block, yet blockNumberOrTagOrHash representing latest or pending: returnLatest=${returnLatest}, blockNumberOrTagOrHash=${blockNumberOrTagOrHash}`
         );
       }
       return null;
@@ -255,7 +270,7 @@ export class CommonService implements ICommonService {
     if (blockNumberOrTagOrHash === EthImpl.emptyHex) {
       if (this.logger.isLevelEnabled('debug')) {
         this.logger.debug(
-          `${requestDetails.formattedRequestId} Invalid input detected in getHistoricalBlockResponse(): blockNumberOrTagOrHash=${blockNumberOrTagOrHash}.`,
+          `${requestDetails.formattedRequestId} Invalid input detected in getHistoricalBlockResponse(): blockNumberOrTagOrHash=${blockNumberOrTagOrHash}.`
         );
       }
       return null;
@@ -295,15 +310,15 @@ export class CommonService implements ICommonService {
     const blockNumberCached = await this.cacheService.getAsync(
       cacheKey,
       CommonService.latestBlockNumber,
-      requestDetails,
+      requestDetails
     );
 
     if (blockNumberCached) {
       if (this.logger.isLevelEnabled('trace')) {
         this.logger.trace(
           `${requestDetails.formattedRequestId} returning cached value ${cacheKey}:${JSON.stringify(
-            blockNumberCached,
-          )}`,
+            blockNumberCached
+          )}`
         );
       }
       return blockNumberCached;
@@ -319,7 +334,7 @@ export class CommonService implements ICommonService {
         currentBlock,
         CommonService.latestBlockNumber,
         requestDetails,
-        this.ethBlockNumberCacheTtlMs,
+        this.ethBlockNumberCacheTtlMs
       );
 
       return currentBlock;
@@ -348,7 +363,7 @@ export class CommonService implements ICommonService {
   public async validateBlockHashAndAddTimestampToParams(
     params: any,
     blockHash: string,
-    requestDetails: RequestDetails,
+    requestDetails: RequestDetails
   ) {
     try {
       const block = await this.mirrorNodeClient.getBlock(blockHash, requestDetails);
@@ -381,7 +396,7 @@ export class CommonService implements ICommonService {
   public async getLogsByAddress(address: string | string[], params: any, requestDetails: RequestDetails) {
     const addresses = Array.isArray(address) ? address : [address];
     const logPromises = addresses.map((addr) =>
-      this.mirrorNodeClient.getContractResultsLogsByAddress(addr, requestDetails, params, undefined),
+      this.mirrorNodeClient.getContractResultsLogsByAddress(addr, requestDetails, params, undefined)
     );
 
     const logResults = await Promise.all(logPromises);
@@ -396,7 +411,7 @@ export class CommonService implements ICommonService {
   public async getLogsWithParams(
     address: string | string[] | null,
     params: any,
-    requestDetails: RequestDetails,
+    requestDetails: RequestDetails
   ): Promise<Log[]> {
     const EMPTY_RESPONSE = [];
 
@@ -423,8 +438,8 @@ export class CommonService implements ICommonService {
           removed: false,
           topics: log.topics,
           transactionHash: toHash32(log.transaction_hash),
-          transactionIndex: numberTo0x(log.transaction_index),
-        }),
+          transactionIndex: numberTo0x(log.transaction_index)
+        })
       );
     }
 
@@ -437,7 +452,7 @@ export class CommonService implements ICommonService {
     toBlock: string | 'latest',
     address: string | string[] | null,
     topics: any[] | null,
-    requestDetails: RequestDetails,
+    requestDetails: RequestDetails
   ): Promise<Log[]> {
     const EMPTY_RESPONSE = [];
     const params: any = {};
@@ -460,7 +475,7 @@ export class CommonService implements ICommonService {
   async resolveEvmAddress(
     address: string,
     requestDetails: RequestDetails,
-    searchableTypes = [constants.TYPE_CONTRACT, constants.TYPE_TOKEN, constants.TYPE_ACCOUNT],
+    searchableTypes = [constants.TYPE_CONTRACT, constants.TYPE_TOKEN, constants.TYPE_ACCOUNT]
   ): Promise<string> {
     if (!address) return address;
 
@@ -469,7 +484,7 @@ export class CommonService implements ICommonService {
       EthImpl.ethGetCode,
       requestDetails,
       searchableTypes,
-      0,
+      0
     );
     let resolvedAddress = address;
     if (
@@ -514,7 +529,108 @@ export class CommonService implements ICommonService {
       // for legacy EIP155 with tx.chainId=0x0, mirror-node will return a '0x' (EMPTY_HEX) value for contract result's chain_id
       //   which is incompatibile with certain tools (i.e. foundry). By setting this field, chainId, to undefined, the end jsonrpc
       //   object will leave out this field, which is the proper behavior for other tools to be compatible with.
-      chainId: cr.chain_id === '0x' ? undefined : cr.chain_id,
+      chainId: cr.chain_id === '0x' ? undefined : cr.chain_id
     };
   };
+
+  /**
+   * Returns the current network fee in denominated in weibars.
+   *
+   * @param callerName
+   * @param requestDetails
+   * @param timestamp
+   */
+  public async getFeeWeibars(callerName: string, requestDetails: RequestDetails, timestamp?: string): Promise<number> {
+    let networkFees;
+
+    try {
+      networkFees = await this.mirrorNodeClient.getNetworkFees(requestDetails, timestamp, undefined);
+    } catch (e: any) {
+      this.logger.warn(
+        e,
+        `${requestDetails.formattedRequestId} Mirror Node threw an error while retrieving fees. Fallback to consensus node.`
+      );
+    }
+
+    if (_.isNil(networkFees)) {
+      if (this.logger.isLevelEnabled('debug')) {
+        this.logger.debug(
+          `${requestDetails.formattedRequestId} Mirror Node returned no network fees. Fallback to consensus node.`
+        );
+      }
+      networkFees = {
+        fees: [
+          {
+            gas: await this.hapiService.getSDKClient().getTinyBarGasFee(callerName, requestDetails),
+            transaction_type: EthImpl.ethTxType
+          }
+        ]
+      };
+    }
+
+    if (networkFees && Array.isArray(networkFees.fees)) {
+      const txFee = networkFees.fees.find(({ transaction_type }) => transaction_type === EthImpl.ethTxType);
+      if (txFee?.gas) {
+        // convert tinyBars into weiBars
+        const weibars = Hbar.fromTinybars(txFee.gas).toTinybars().multiply(constants.TINYBAR_TO_WEIBAR_COEF);
+
+        return weibars.toNumber();
+      }
+    }
+
+    throw predefined.COULD_NOT_ESTIMATE_GAS_PRICE;
+  }
+
+  /**
+   * Retrieves the current network gas price in weibars.
+   *
+   * @returns {Promise<string>} The current gas price in weibars as a hexadecimal string.
+   * @throws Will throw an error if unable to retrieve the gas price.
+   * @param requestDetails
+   */
+  public async gasPrice(requestDetails: RequestDetails): Promise<string> {
+    if (this.logger.isLevelEnabled('trace')) {
+      this.logger.trace(`${requestDetails.formattedRequestId} eth_gasPrice`);
+    }
+    try {
+      let gasPrice: number | undefined = await this.cacheService.getAsync(
+        constants.CACHE_KEY.GAS_PRICE,
+        EthImpl.ethGasPrice,
+        requestDetails
+      );
+
+      if (!gasPrice) {
+        gasPrice = Utils.addPercentageBufferToGasPrice(await this.getFeeWeibars(EthImpl.ethGasPrice, requestDetails));
+
+        await this.cacheService.set(
+          constants.CACHE_KEY.GAS_PRICE,
+          gasPrice,
+          EthImpl.ethGasPrice,
+          requestDetails,
+          this.ethGasPriceCacheTtlMs
+        );
+      }
+
+      return numberTo0x(gasPrice);
+    } catch (error) {
+      throw this.genericErrorHandler(error, `${requestDetails.formattedRequestId} Failed to retrieve gasPrice`);
+    }
+  }
+
+  /**
+   * Translates a block tag into a number. 'latest', 'pending', and null are the most recent block, 'earliest' is 0, numbers become numbers.
+   *
+   * @param tag null, a number, or 'latest', 'pending', or 'earliest'
+   * @param requestDetails
+   * @private
+   */
+  public async translateBlockTag(tag: string | null, requestDetails: RequestDetails): Promise<number> {
+    if (this.blockTagIsLatestOrPending(tag)) {
+      return Number(await this.getLatestBlockNumber(requestDetails));
+    } else if (tag === EthImpl.blockEarliest) {
+      return 0;
+    } else {
+      return Number(tag);
+    }
+  }
 }
